requirement.txt
langchain
langchain-core
langchain-community
langchain-mistralai

openai
python-dotenv

pydantic
tiktoken

faiss-cpu
chromadb

sentence-transformers

streamlit

unstructured
pdfplumber
pypdf

requests
numpy
pandas
